# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Que Lo Que delivery platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user (role-based)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, picked_up, delivered, cancelled]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (business/driver/admin)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderId}/rate:
    post:
      operationId: rateOrder
      tags: [orders]
      summary: Rate a completed order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateOrderBody"
      responses:
        "200":
          description: Rating submitted

  # DRIVERS
...
  /customer/addresses:
    get:
      operationId: listMyAddresses
      tags: [customers]
      summary: Get saved addresses for the logged-in customer
      responses:
        "200":
          description: List of saved addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserAddress"
    post:
      operationId: saveAddress
      tags: [customers]
      summary: Save a new delivery address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAddressBody"
      responses:
        "201":
          description: Address saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAddress"
  /customer/addresses/{id}:
    delete:
      operationId: deleteAddress
      tags: [customers]
      summary: Delete a saved address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # STATS
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [customer, driver, business, admin]
        isBanned:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isBanned
        - createdAt
...
    Business:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [food, supermarket, pharmacy, liquor]
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        imageUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        isActive:
          type: boolean
        rating:
          type: number
        totalOrders:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - name
        - category
        - address
        - isActive
        - createdAt
...
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        businessId:
          type: integer
        driverId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, picked_up, delivered, cancelled]
        totalAmount:
          type: number
        deliveryFee:
          type: number
        commission:
          type: number
        driverEarnings:
          type: number
        tip:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card]
        isPaid:
          type: boolean
        deliveryAddress:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        business:
          $ref: "#/components/schemas/Business"
        customer:
          $ref: "#/components/schemas/User"
        driver:
          $ref: "#/components/schemas/Driver"
        customerRating:
          type: integer
        driverRating:
          type: integer
        businessRating:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - customerId
        - businessId
        - status
        - totalAmount
        - deliveryFee
        - paymentMethod
        - isPaid
        - deliveryAddress
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - orderId
        - productName
        - quantity
        - price
    CreateOrderBody:
      type: object
      properties:
        businessId:
          type: integer
        paymentMethod:
          type: string
          enum: [cash, card]
        deliveryAddress:
          type: string
        notes:
          type: string
        tip:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - businessId
        - paymentMethod
        - deliveryAddress
        - items
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [accepted, picked_up, delivered, cancelled]
      required:
        - status
    RateOrderBody:
      type: object
      properties:
        driverRating:
          type: integer
        businessRating:
          type: integer
        comment:
          type: string
      required:
        - driverRating
        - businessRating
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        isOnline:
          type: boolean
        isLocked:
          type: boolean
        rating:
          type: number
        acceptanceRate:
          type: number
        currentLat:
          type: number
        currentLng:
          type: number
        cashBalance:
          type: number
        walletBalance:
          type: number
        totalDeliveries:
          type: integer
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - userId
        - isOnline
        - isLocked
        - rating
        - acceptanceRate
        - cashBalance
        - walletBalance
        - totalDeliveries
        - createdAt
...
    UserAddress:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        label:
          type: string
        address:
          type: string
        isDefault:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - label
        - address
        - isDefault
        - createdAt
    SaveAddressBody:
      type: object
      properties:
        label:
          type: string
        address:
          type: string
        isDefault:
          type: boolean
      required:
        - label
        - address