# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Que Lo Que delivery platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";
import type { OrderItemInput } from "./orderItemInput";

export interface CreateOrderBody {
  businessId: number;
  paymentMethod: CreateOrderBodyPaymentMethod;
  deliveryAddress: string;
  notes?: string;
  tip?: number;
  items: OrderItemInput[];
}