# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Que Lo Que delivery platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Business } from "./business";
import type { Driver } from "./driver";
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderStatus } from "./orderStatus";
import type { User } from "./user";

export interface Order {
  id: number;
  customerId: number;
  businessId: number;
  driverId?: number;
  status: OrderStatus;
  totalAmount: number;
  deliveryFee: number;
  commission?: number;
  driverEarnings?: number;
  tip?: number;
  paymentMethod: OrderPaymentMethod;
  isPaid: boolean;
  deliveryAddress: string;
  notes?: string;
  items?: OrderItem[];
  business?: Business;
  customer?: User;
  driver?: Driver;
  customerRating?: number;
  driverRating?: number;
  businessRating?: number;
  createdAt: string;
}