# lib/api-zod/src/generated/types/userAddress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Que Lo Que delivery platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UserAddress {
  id: number;
  userId: number;
  label: string;
  address: string;
  isDefault: boolean;
  createdAt: Date;
}